feat: 사용자 가입 기능을 UserController로 분리
- 사용자 가입 로직을 UserController.js로 이동
- users.js에서 가입 라우터를 join 함수로 연결
