+++ backend/routes/users.js
@@ -1,22 +1,9 @@
 const express = require("express");
 const router = express.Router();
-const conn = require("../mariadb");
-const { StatusCodes } = require("http-status-codes");
+const { join } = require("../controller/UserController");
 
 router.use(express.json());
-router.post("/join", (req, res) => {
-  const { email, password } = req.body;
-  let sql = "INSERT INTO users (email, password) VALUES (?, ?)";
-  let values = [email, password];
-
-  conn.query(sql, values, (err, results) => {
-    if (err) {
-      console.log(err);
-      res.status(StatusCodes.BAD_REQUEST).end();
-    }
-    return res.status(StatusCodes.CREATED).json(results);
-  });
-});
+router.post("/join", join);
 
 router.post("/login", (req, res) => {
   res.json({ message: "로그인" });
